pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "git-safety-agent"
version = "0.1.0"
description = "Git Safety Agent"
readme = "README.md"
requires-python = ">=3.10"
authors = [
  {name = "git-safety-agent", email = "dev@example.com"}
]
license = {text = "MIT"}
dependencies = [
  "pydantic>=2.5",
  "fastapi>=0.110",
  "uvicorn>=0.23",
  "streamlit>=1.32",
  "httpx>=0.27",
  "PyYAML>=6.0",
  "zai-sdk>=0.2.0",
  "langchain-core>=0.2",
  "langchain-community>=0.2",
  "langchain-text-splitters>=0.2",
  "faiss-cpu>=1.7",
  "numpy>=1.26",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
]

[project.scripts]
gsa = "gsa.cli:main"

[tool.setuptools.packages.find]
where = ["src"]
